api/requirements.txt
blinker==1.8.2
click==8.1.7
Flask==3.0.3  
gunicorn==23.0.0
importlib_metadata==8.5.0
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==3.0.1
packaging==24.1
Werkzeug==3.0.4
zipp==3.20.2

# Dependencias adicionales
Flask-JWT-Extended==4.6.0
Flask-Migrate==3.1.0
Flask-SQLAlchemy==3.0.0
